refactor: Refactor fee-related types (#3121)
- Moves basic logic for fee types close to the type declaration (i.e.,
to `zksync_types`).
- Unifies `PubdataSendingMode` and `PubdataDA`.
- Removes dependency of `zksync_types` on `zksync_config`.

This allows external apps depending on fee types (e.g., the test node)
to not have behemoths like `zksync_dal` in their dependency graph.

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.
