+++ core/lib/env_config/src/eth_sender.rs
@@ -41,7 +41,8 @@ impl FromEnv for GasAdjusterConfig {
 
 #[cfg(test)]
 mod tests {
-    use zksync_config::configs::eth_sender::{ProofSendingMode, PubdataSendingMode};
+    use zksync_basic_types::pubdata_da::PubdataSendingMode;
+    use zksync_config::configs::eth_sender::ProofSendingMode;
 
     use super::*;
     use crate::test_utils::{hash, EnvMutex};
