+++ core/lib/types/src/lib.rs
@@ -44,7 +44,6 @@ pub mod l2;
 pub mod l2_to_l1_log;
 pub mod priority_op_onchain_data;
 pub mod protocol_upgrade;
-pub mod pubdata_da;
 pub mod snapshots;
 pub mod storage;
 pub mod system_contracts;
