+++ core/node/fee_model/src/l1_gas_price/gas_adjuster/mod.rs
@@ -6,9 +6,12 @@ use std::{
 };
 
 use tokio::sync::watch;
-use zksync_config::{configs::eth_sender::PubdataSendingMode, GasAdjusterConfig};
+use zksync_config::GasAdjusterConfig;
 use zksync_eth_client::EthFeeInterface;
-use zksync_types::{commitment::L1BatchCommitmentMode, L1_GAS_PER_PUBDATA_BYTE, U256};
+use zksync_types::{
+    commitment::L1BatchCommitmentMode, pubdata_da::PubdataSendingMode, L1_GAS_PER_PUBDATA_BYTE,
+    U256,
+};
 use zksync_web3_decl::client::{DynClient, L1, L2};
 
 use self::metrics::METRICS;
