+++ core/node/node_framework/src/implementations/layers/gas_adjuster.rs
@@ -1,8 +1,9 @@
 use std::sync::Arc;
 
 use anyhow::Context;
-use zksync_config::{configs::eth_sender::PubdataSendingMode, GasAdjusterConfig, GenesisConfig};
+use zksync_config::{GasAdjusterConfig, GenesisConfig};
 use zksync_node_fee_model::l1_gas_price::GasAdjuster;
+use zksync_types::pubdata_da::PubdataSendingMode;
 
 use crate::{
     implementations::resources::{
